refactor(hamcrest): Change packages for the test classes to corresponding directories
