+++ src/main/java/org/mockbukkit/mockbukkit/world/WorldMock.java
@@ -1,10 +1,6 @@
 package org.mockbukkit.mockbukkit.world;
 
 import org.mockbukkit.mockbukkit.AsyncCatcher;
-import org.mockbukkit.mockbukkit.world.chunk.ChunkMock;
-import org.mockbukkit.mockbukkit.world.chunk.ChunkSnapshotMock;
-import org.mockbukkit.mockbukkit.world.coordinate.ChunkCoordinate;
-import org.mockbukkit.mockbukkit.world.coordinate.Coordinate;
 import org.mockbukkit.mockbukkit.MockBukkit;
 import org.mockbukkit.mockbukkit.ServerMock;
 import org.mockbukkit.mockbukkit.block.BlockMock;
