+++ src/test/java/org/mockbukkit/mockbukkit/block/BlockMockTest.java
@@ -14,9 +14,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.mockbukkit.mockbukkit.world.coordinate.ChunkCoordinate;
-import org.mockbukkit.mockbukkit.world.chunk.ChunkMock;
-import org.mockbukkit.mockbukkit.world.coordinate.Coordinate;
+import org.mockbukkit.mockbukkit.world.ChunkCoordinate;
+import org.mockbukkit.mockbukkit.world.ChunkMock;
+import org.mockbukkit.mockbukkit.world.Coordinate;
 import org.mockbukkit.mockbukkit.MockBukkitExtension;
 import org.mockbukkit.mockbukkit.world.WorldMock;
 import org.mockbukkit.mockbukkit.block.data.BlockDataMock;
