+++ src/test/java/org/mockbukkit/mockbukkit/plugin/PermissionManagerMockTest.java
@@ -1,5 +1,9 @@
-package org.mockbukkit.mockbukkit;
+package org.mockbukkit.mockbukkit.plugin;
 
+import org.mockbukkit.mockbukkit.MockBukkit;
+import org.mockbukkit.mockbukkit.MockBukkitExtension;
+import org.mockbukkit.mockbukkit.MockBukkitInject;
+import org.mockbukkit.mockbukkit.ServerMock;
 import org.mockbukkit.mockbukkit.entity.BatMock;
 import org.bukkit.entity.Player;
 import org.bukkit.permissions.Permission;
@@ -10,7 +14,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.mockbukkit.mockbukkit.plugin.PermissionManagerMock;
 
 import java.util.List;
 import java.util.UUID;
