+++ src/test/java/org/mockbukkit/mockbukkit/plugin/PluginManagerMockTest.java
@@ -22,7 +22,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.mockbukkit.mockbukkit.MockBukkit;
 import org.mockbukkit.mockbukkit.ServerMock;
-import org.mockbukkit.mockbukkit.TestPlugin;
 import org.mockbukkit.mockbukkit.exception.EventHandlerException;
 
 import java.lang.reflect.Field;
