+++ src/test/java/org/mockbukkit/mockbukkit/plugin/TestPlugin.java
