+++ src/test/java/org/mockbukkit/mockbukkit/registry/RegistryMockTest.java
