+++ src/test/java/org/mockbukkit/mockbukkit/sound/MusicInstrumentMockTest.java
@@ -1,11 +1,11 @@
-package org.mockbukkit.mockbukkit;
+package org.mockbukkit.mockbukkit.sound;
 
 import com.google.gson.JsonObject;
 import org.bukkit.NamespacedKey;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockbukkit.mockbukkit.sound.MusicInstrumentMock;
+import org.mockbukkit.mockbukkit.MockBukkitExtension;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
