+++ src/test/java/org/mockbukkit/mockbukkit/util/UnsafeValuesTest.java
@@ -1,5 +1,9 @@
-package org.mockbukkit.mockbukkit;
+package org.mockbukkit.mockbukkit.util;
 
+import org.mockbukkit.mockbukkit.MockBukkit;
+import org.mockbukkit.mockbukkit.MockBukkitExtension;
+import org.mockbukkit.mockbukkit.MockBukkitInject;
+import org.mockbukkit.mockbukkit.ServerMock;
 import org.mockbukkit.mockbukkit.exception.UnimplementedOperationException;
 import org.mockbukkit.mockbukkit.inventory.ItemStackMock;
 import com.google.gson.JsonElement;
