+++ src/test/resources/empty_plugin.yml
@@ -1,3 +1,3 @@
 name: SecondTestPlugin
-main: org.mockbukkit.mockbukkit.EmptyPlugin
+main: org.mockbukkit.mockbukkit.plugin.EmptyPlugin
 version: 0.1.0
