+++ src/test/resources/second_plugin.yml
@@ -1,3 +1,3 @@
 name: SecondTestPlugin
-main: org.mockbukkit.mockbukkit.SecondTestPlugin
+main: org.mockbukkit.mockbukkit.plugin.SecondTestPlugin
 version: 0.1.0
